changed price to int

+++ createSQLtable.sql
@@ -5,7 +5,7 @@ tables and have a key that auto increments*/
 CREATE TABLE `my_products` (
   `unique_id` int(11) NOT NULL,
   `product_name` varchar(255) NOT NULL,
-  `product_price` double NOT NULL,
+  `product_price` int NOT NULL,
   `product_desc` text NOT NULL,
   `product_short_desc` varchar(255) NOT NULL,
   `product_img` text NOT NULL,
